// src/models/enums/shared.enum.ts
export enum DownloadType {
  Docs = 'doc',
  PDF = 'pdf',
  Text = 'text',
}

export enum EHttpStatusCode {
  Accepted = 202,
  AlreadyReported = 208,
  BadGateway = 502,
  BadRequest = 400,
  Conflict = 409,
  Continue = 100,
  Created = 201,
  EarlyHints = 103,
  ExpectationFailed = 417,
  FailedDependency = 424,
  Forbidden = 403,
  Found = 302,
  GatewayTimeout = 504,
  Gone = 410,
  HttpVersionNotSupported = 505,
  ImATeapot = 418,
  ImUsed = 226,
  InsufficientStorage = 507,
  InternalServerError = 500,
  LengthRequired = 411,
  Locked = 423,
  LoopDetected = 508,
  MethodNotAllowed = 405,
  MisdirectedRequest = 421,
  MovedPermanently = 301,
  MultipleChoices = 300,
  MultiStatus = 207,
  NetworkAuthenticationRequired = 511,
  NoContent = 204,
  NonAuthoritativeInformation = 203,
  NotAcceptable = 406,
  NotExtended = 510,
  NotFound = 404,
  NotImplemented = 501,
  NotModified = 304,
  Ok = 200,
  PartialContent = 206,
  PayloadTooLarge = 413,
  PaymentRequired = 402,
  PermanentRedirect = 308,
  PreconditionFailed = 412,
  PreconditionRequired = 428,
  Processing = 102,
  ProxyAuthenticationRequired = 407,
  RangeNotSatisfiable = 416,
  RequestHeaderFieldsTooLarge = 431,
  RequestTimeout = 408,
  ResetContent = 205,
  SeeOther = 303,
  ServiceUnavailable = 503,
  SwitchingProtocols = 101,
  TemporaryRedirect = 307,
  TooEarly = 425,
  TooManyRequests = 429,
  Unauthorized = 401,
  UnavailableForLegalReasons = 451,
  UnprocessableEntity = 422,
  UnsupportedMediaType = 415,
  Unused = 306,
  UpgradeRequired = 426,
  UriTooLong = 414,
  UseProxy = 305,
  VariantAlsoNegotiates = 506,
}

export enum ELanguageCode {
  English = 'en',
  Japanese = 'ja',
  Vietnamese = 'vi',
}

export enum EMessage {
  Error = 'error',
  Info = 'info',
  Success = 'success',
  Warning = 'warning',
}

export enum ERegionalLocale {
  EnglishUS = 'en-US',
  JapaneseJP = 'ja-JP',
  VietnameseVN = 'vi-VN',
}

export enum EResponseStatus {
  Failure = 'failure',
  Success = 'success',
}

export enum EToast {
  Error = 'error',
  Info = 'info',
  Success = 'success',
  Warning = 'warning',
}